Update grocery.html
I created a table with the user's grocery list by using the bootstrap container-5 class. I added form using the form-class, class in bootstrap. I created an "Add to Kitchen" button that pops-up the form mentioned above for users to insert the food item and food group and check off whether the food item was bought or added to kitchen. To help users with the food group section, I added a mini note underneath that field describing what the food group options are. The "Add to Kitchen" button brings the form as a pop-up by using the modal plugin that comes with bootstrap.