Kitchen Page CSS Animations
Float shadow CSS animation was applied to the food groups. Grow shadow CSS animation was applied to rows in "My Kitchen" table. I adjusted the size the rows would grow when hovered over by changing the transform scale from 1.1 to 1.025 on lines 263 and 264. I added the grow shadow id on line 12 in the kitchen.js file. Curl top right CSS animation was applied to "Add Item" button at bottom of page. I changed the background color after the curl effect is applied to a light orange color (#E99D3E) on line 284. I also decreased the width and height of the curl on lines 294 and 295.