Grocery and Kitchen CSS Animation
Clarification for Kitchen Curl Top Right Animation: I changed the background color to light orange (#f0b670) on line 286.

Grow shadow CSS animation was applied to rows in "My Grocery List" table. I adjusted the size the rows would grow when hovered over by changing the transform scale from 1.1 to 1.025 on lines 263 and 264. Since I wanted the background color of the top-right corner to be gray after being hovered over and curling, I created a new bhvr-curl-top-right id with the CSS to do just that on line 317. I also decreased the width and height of the curl on lines 325 and 326 from 25px to 10px.