Donation Page CSS Animations
I added the orange underline to the "Donation List" title on the donationlist page just as I did to the "Grocery List" title on the grocery page. I applied the hover grow shadow CSS animation to the rows in the "Donation List" table and sign up and login buttons on the top right hand corner of the page. I also applied the black curl top right CSS animation to the "Add to Donation List" button on the bottom right hand side of the page.